+++ routes/routes.py
@@ -12,6 +12,7 @@ from routes.content_routes import router as content_router
 from routes.api_routes import router as api_router
 from routes.news_routes import router as news_admin_router
 from routes.news_content_routes import router as news_content_router
+from routes.modules_routes import router as modules_router
 
 
 router = APIRouter()
@@ -23,3 +24,4 @@ router.include_router(content_router)
 router.include_router(api_router)
 router.include_router(news_admin_router)
 router.include_router(news_content_router)
+router.include_router(modules_router)